+++ src/shared_math/other.rs
@@ -12,12 +12,12 @@ const fn num_bits<T>() -> u64 {
     std::mem::size_of::<T>() as u64 * 8
 }
 
-pub fn log_2_floor(x: u64) -> u64 {
+pub fn log_2_floor(x: u128) -> u64 {
     assert!(x > 0);
     num_bits::<u64>() - x.leading_zeros() as u64 - 1
 }
 
-pub fn log_2_ceil(x: u64) -> u64 {
+pub fn log_2_ceil(x: u128) -> u64 {
     if is_power_of_two(x) {
         log_2_floor(x)
     } else {
@@ -42,7 +42,7 @@ where
 
 /// Round up to the nearest power of 2
 pub fn roundup_npo2(x: u64) -> u64 {
-    1 << log_2_ceil(x)
+    1 << log_2_ceil(x as u128)
 }
 
 #[inline]
@@ -137,7 +137,7 @@ pub fn get_node_count_in_complete_binary_tree(leaf_count: usize) -> usize {
 pub fn get_height_of_complete_binary_tree(leaf_count: usize) -> usize {
     assert!(is_power_of_two(leaf_count));
     // nightly: leaves_count.log2() + 1
-    log_2_floor(leaf_count as u64) as usize
+    log_2_floor(leaf_count as u128) as usize
 }
 
 /// Count bits required to represent an unsigned number. Inspired by and similar to
@@ -164,18 +164,18 @@ mod test_other {
         assert_eq!(4, log_2_floor(16));
         assert_eq!(1, log_2_floor(2));
         assert_eq!(0, log_2_floor(1));
-        assert_eq!(40, log_2_floor(2u64.pow(40)));
-        assert_eq!(40, log_2_floor(2u64.pow(40) + 1));
-        assert_eq!(40, log_2_floor(2u64.pow(40) + 456456));
+        assert_eq!(40, log_2_floor(2u128.pow(40)));
+        assert_eq!(40, log_2_floor(2u128.pow(40) + 1));
+        assert_eq!(40, log_2_floor(2u128.pow(40) + 456456));
         assert_eq!(4, log_2_ceil(16));
         assert_eq!(5, log_2_ceil(17));
         assert_eq!(5, log_2_ceil(18));
         assert_eq!(5, log_2_ceil(19));
         assert_eq!(1, log_2_ceil(2));
         assert_eq!(0, log_2_ceil(1));
-        assert_eq!(40, log_2_ceil(2u64.pow(40)));
-        assert_eq!(41, log_2_ceil(2u64.pow(40) + 1));
-        assert_eq!(41, log_2_ceil(2u64.pow(40) + 456456));
+        assert_eq!(40, log_2_ceil(2u128.pow(40)));
+        assert_eq!(41, log_2_ceil(2u128.pow(40) + 1));
+        assert_eq!(41, log_2_ceil(2u128.pow(40) + 456456));
     }
 
     #[test]
